+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# [Unreleased](https://github.com/MNM-DTU-DZ2/GreenEduMap-DTUDZ/compare/v0.2.0...f9a73167a897dfe8a17db3e74525bdb2728291c3) (2025-12-02)
+
+
+### Bug Fixes
+
+* **ci:** add -r 0 flag to regenerate full changelog with unreleased ([00e7991](https://github.com/MNM-DTU-DZ2/GreenEduMap-DTUDZ/commit/00e7991d3036604c6f98ac11344c8c60447bb7b4))
+
+
+### Features
+
+* **auth:** implement  user authentication and profile system ([f9a7316](https://github.com/MNM-DTU-DZ2/GreenEduMap-DTUDZ/commit/f9a73167a897dfe8a17db3e74525bdb2728291c3)), closes [#8](https://github.com/MNM-DTU-DZ2/GreenEduMap-DTUDZ/issues/8)
+* **reviews,testing:** implement user reviews system and E2E testing ([f51d710](https://github.com/MNM-DTU-DZ2/GreenEduMap-DTUDZ/commit/f51d7108708cbe76f8c6b6ab087917ae9132b419))
+
+
+
 # [0.2.0](https://github.com/MNM-DTU-DZ2/GreenEduMap-DTUDZ/compare/v0.1.0...v0.2.0) (2025-12-01)